chore(folder structure): adjusted folder structure
